% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_format_function.R
\name{check_data_format}
\alias{check_data_format}
\title{Reads csv with meaningful error message}
\usage{
check_data_format(file)
}
\arguments{
\item{file}{A filename as a character vector of length 1}
}
\value{
A data frame containing data from the .csv or .xlsx file
}
\description{
Using the my_read_csv_function and my_read_xlsx_function to check the data format to test if R can open data.
}
