% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_na_function.R
\name{contains.na}
\alias{contains.na}
\title{Locates NAs within the dataset}
\usage{
contains.na(x)
}
\arguments{
\item{fn}{A filename as a character vector of length 1}

\item{x}{A dataframe}
}
\value{
Either TRUE or FALSE if there is/are NAs in the dataset, and where ther are
}
\description{
Function determines if and where NAs are in the dataset.
}
