% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_read_xlsx_function.R
\name{my_read_xlsx_function}
\alias{my_read_xlsx_function}
\title{Reads xlsx files with meaningful error message}
\usage{
my_read_xlsx_function(file)
}
\arguments{
\item{file}{A filename as a character vector of length 1}
}
\value{
A data frame representing the .xlsx file
}
\description{
Function to read both .csv and .xlsx files into R.
}
