% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_columns_function.R
\name{numeric_columns}
\alias{numeric_columns}
\title{Determines if columns contain valid numeric values and where non-numeric values are located}
\usage{
numeric_columns(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A data frame representing the data from .csv or .xlsx file
}
\description{
Function which determines if columns contain valid numeric values and, if not, where non-numeric values are located in the data frame.
}
