% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_survey.R
\name{read_survey}
\alias{read_survey}
\title{Read survey data}
\usage{
read_survey(fn)
}
\arguments{
\item{fn}{A filename as a character vector of length 1}

\item{sheet}{For .xlsx files, the name of the spreadsheet in which the data are stored}
}
\value{
A data frame containing data from the .csv or .xlsx file
}
\description{
Interprets the extension of a filename for survey data and calls the appropriate function to read the file.
}
\details{
Currently this can read .xslx or .csv files which must be formatted according to the vignette. Any file not ending in .xlsx tries to read a .csv
}
