% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_survey_results_csv.R
\name{read_survey_results_csv}
\alias{read_survey_results_csv}
\title{Reads csv files with meaningful error message}
\usage{
read_survey_results_csv(file)
}
\arguments{
\item{file}{A filename as a character vector of length 1}
}
\value{
A data frame representing the .csv file
}
\description{
Function to read both .csv and .xlsx files into R.
}
