% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_survey_results_xlsx.R
\name{read_survey_results_xlsx}
\alias{read_survey_results_xlsx}
\title{Reads xlsx files with meaningful error message}
\usage{
read_survey_results_xlsx(file)
}
\arguments{
\item{file}{A filename as a character vector of length 1}
}
\value{
A data frame representing the .xlsx file
}
\description{
Function to read both .csv and .xlsx files into R.
}
