% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach_anxiety.R
\name{teach_anxiety}
\alias{teach_anxiety}
\title{Function to calculate Teaching Anxiety Score from Parson's (1973) survey}
\usage{
teach_anxiety(x)
}
\arguments{
\item{x}{A dataframe. NAs are allowed. Use contains_na_function to detect where NAs are located.}
}
\value{
A data frame with the Teaching Anxiety score as the last column.
}
\description{
Function does TWO things: (1) reverse scores 14 questions, and (2) sums all columns for each row/participant to calculate an Anxiety Score.
}
\details{
Function to calculate Teaching Anxiety score.
}
